[MCUX-73084] Fix send issue on USB_EthernetIfOutPut.
In the case that p->tot_len in the RNDIS transmit portion of
USB_EthernetIfOutPut() is not equal to p->len (i.e. when the long frame
is split into multiple parts), since the messageBuffer & dataBuffer
share the same buffer passed into USB_HostRndisSendDataMsg(), the
implement code of USB_HostRndisSendDataMsg() will organize the packets,
that it will change the contents in messageBuffer, and this also
modifies the Ethernet frames in the dataBuffer.

This patch have fixed the issue that the behavior causes the USB Host to
send an incorrect Ethernet frame in this scenario, which can seriously
affect the behavior of the app.

Signed-off-by: Dv Alan <zhangyang.shen@nxp.com>