[MCUX-79993] Fix ethernet transmission issue on USB.
This patch fix the transmission issue on USB that one ethernet packet
which consists of multiple pbuf.

In some scenarios (where LwIP uses multiple pbufs to represent a single
packet) such as TLS handshack state or other, the whole data of the
packet is distributed into multiple pbuf payloads, that means this
complete data content is not contiguous in memory. Using
pbuf_copy_partial() to ensure that packet data is contiguous in specific
buffer.

Signed-off-by: Dv Alan <zhangyang.shen@nxp.com>