[WSW-60641] relocate lwiperf_txbuf_const to quick access data section
In zero-copy cases, DMA will copy from lwiperf_txbuf_const
to peripherals.
But as a const variable, it is in flash, which is much slower
for DMA to access.
Eg. wifi SDIO zero copy cases TCP TX iperf drops from 89Mbps
to 76Mbps.

So relocate it to quick access data section to optimize lwiperf
performance.

Signed-off-by: Fengming Ye <frank.ye@nxp.com>